requirements.txt
black==22.3.0; python_full_version >= "3.6.2"
brotli==1.0.9; python_version >= "3.6"
click==8.1.3; python_version >= "3.7" and python_full_version >= "3.6.2"
colorama==0.4.4; platform_system == "Windows" and python_version >= "3.7" and python_full_version >= "3.6.2"
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-renderer==1.9.1
dash-table==5.0.0
dash==2.4.1; python_version >= "3.6"
flask-assets==2.0
flask-compress==1.12; python_version >= "3.6"
flask==2.1.2; python_version >= "3.7"
importlib-metadata==4.11.4; python_version < "3.10" and python_version >= "3.7"
isort==5.10.1; python_full_version >= "3.6.1" and python_version < "4.0"
itsdangerous==2.1.2; python_version >= "3.7"
jinja2==3.1.2; python_version >= "3.7"
markupsafe==2.1.1; python_version >= "3.7"
mypy-extensions==0.4.3; python_full_version >= "3.6.2"
numpy==1.23.5
pandas==1.4.2; python_version >= "3.8"
pathspec==0.9.0; python_full_version >= "3.6.2"
platformdirs==2.5.2; python_version >= "3.7" and python_full_version >= "3.6.2"
plotly==5.8.0; python_version >= "3.6"
python-dateutil==2.8.2; python_version >= "3.8" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.8"
python-dotenv==0.20.0; python_version >= "3.5"
pytz==2022.1; python_version >= "3.8"
six==1.16.0; python_version >= "3.8" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.8"
tenacity==8.0.1; python_version >= "3.6"
tomli==2.0.1; python_version < "3.11" and python_full_version >= "3.6.2" and python_version >= "3.7"
typing-extensions==4.2.0; python_version < "3.10" and python_full_version >= "3.6.2" and python_version >= "3.7"
webassets==2.0
werkzeug==2.1.2; python_version >= "3.7"
zipp==3.8.0; python_version < "3.10" and python_version >= "3.7"
SQLAlchemy==1.4.38
Flask-SQLAlchemy==2.5.1
Flask-Login==0.6.2
PyMySQL==1.0.2
cloud-sql-python-connector==1.1.0
Flask-Login==0.6.2
psycopg2-binary
#mysqlclient
flask_wtf
flask_mail
email_validator
cssmin
#ruamel_yaml==0.11.14

#nuevos requiremintos 
networkx 
scipy
matplotlib
seaborn
#sklearn
scikit-learn
ruamel_yaml==0.17.32
mpld3
Flask-JWT-Extended
gunicorn>=20.1.0
Flask-Dance==6.2.0
Flask-OAuth
urllib3==2.0.2
